third_party/WebKit/LayoutTests/http/tests/streams/chromium/get-reader-byob.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
  const rs = new ReadableStream();
  assert_throws(new TypeError(), () => rs.getReader({mode: 'byob'}),
                'a TypeError should be thrown');
}, 'getReader() with mode "byob" should throw a TypeError');
</script>
